Move pll_frequency to ppl submodule
